Move types to @types directory
